==> 07-conclusion.Rmd <==
# Conclusion

Based on teh analysis of the number of collisions in New York City from January to October 2021, we discovered that **Brooklyn** and **Queens** are the top-tier boroughs with the most accidents. Among the accidents, most precincts in **Brooklyn** and **Manhattan** have the highest rate of severe collisions. **Staten Island** is the borough that is relatively safe to drive or walk due to its low number of collisions and low severe rate of accidents. **The Bronx** is neutral among the five boroughs.

In terms of the contributing factors of car accidents, we found that **distracted driving** and **wrong-way driving** are particularly problematic in New York City, which is consistent with the national trend. Furthermore, the distribution of causes varies by location. Brooklyn, in particular, has the highest proportion of distracted driving, while Queens has the highest proportion of wrong-way driving, which explains why these two boroughs have the most severe cases in NYC. To help people focus on driving, drivers should commit to driving without using their phones and limit passengers and pets who may cause distractions. City authorities should also ensure that there are no external distractions on the busy streets, such as digital billboards. 

Our source data has certain limitations. Since it only offers the summary statistics of vehicle crash data for the first 10 months of 2021, we do not have any historical data to compare year to year, nor do we have comprehensive data like monthly or daily to gain insights on seasonality shift. If we have daily data, we could combine it with weather data from NYC which may lead to some interesting findings of weather conditions such as rainy and snowy days.

Furthermore, driver inattention was the major cause of driving distraction. If it could be broken down into additional aspects, such as eating, reading, or applying makeup, we would have a better idea of which factors contribute the most to driver's inattention.

Due to the scope of the project, some directions of research are left for future studies. For example,  we discovered that cyclists have a much higher proportion of injured individuals of car collisions in Manhattan. This might be caused that city-bike and food delivery service is more popular in Manhattan, which leads to an increase in total incidents and raises the proportion. Such a proposition requires more data to support the analysis. Moreover, we were not able to obtain data that records the total number of all individuals involved in car collisions. Thus we can not come up with the actual danger/ injury rate of individuals in car collisions. Incorporating the population, cars ownership, bicycle ownership of each borough in NYC would also facilitate the project to carry out more precise and significant research.
